fix(executor/*astore): support all computional int types
All primitive array types whose component type is a computional
int must be able to retrieve the value to store from
any of the computational int types.

The easiest example is a byte array `byte[] b = {10, 11};`.
This cause the JDK compiler to emit code that first pushes
an *int* (via `bipush`) onto the stack,
which then is stored into the *byte* array (via `bastore`).